Add optimisation._geometry._tools module
The 'optimise.py' file was a bit complex, so split out the utility
functions.
